add db read-only mode support

+++ src/Art.EF.Sqlite/SqliteArtifactRegistrationManager.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.EntityFrameworkCore;
+﻿using System.Text;
+using Microsoft.EntityFrameworkCore;
 
 namespace Art.EF.Sqlite;
 
@@ -11,12 +12,27 @@ public class SqliteArtifactRegistrationManager : EFArtifactRegistrationManager<S
     {
         if (factory.UsingInMemory)
         {
-            Context.Database.EnsureDeleted();
-            Context.Database.EnsureCreated();
+            if (!factory._isReadonly)
+            {
+                Context.Database.EnsureDeleted();
+                Context.Database.EnsureCreated();
+            }
         }
         else
         {
-            Context.Database.Migrate();
+            if (factory._isReadonly)
+            {
+                var pendingMigrations = Context.Database.GetPendingMigrations().ToList();
+                if (pendingMigrations.Count > 0)
+                {
+                    string pendingMigrationsStr = new StringBuilder().AppendJoin(", ", pendingMigrations).ToString();
+                    throw new InvalidOperationException($"There are pending migrations that cannot be applied to a read-only database: {pendingMigrationsStr}");
+                }
+            }
+            else
+            {
+                Context.Database.Migrate();
+            }
         }
     }
 
@@ -34,10 +50,11 @@ public class SqliteArtifactRegistrationManager : EFArtifactRegistrationManager<S
     /// Creates a new instance of <see cref="SqliteArtifactRegistrationManager"/> with in-memory Sqlite backing.
     /// </summary>
     /// <param name="inMemory">If true, use in-memory otherwise allow fallback to environment variable.</param>
+    /// <param name="isReadonly">If true, writes to the database are disabled.</param>
     /// <remarks>
     /// Sqlite file backing if environment variable art_ef_sqlite_backing_file is set and <paramref name="inMemory"/> is false, otherwise in-memory Sqlite backing
     /// </remarks>
-    public SqliteArtifactRegistrationManager(bool inMemory) : this(new SqliteArtifactContextFactory(inMemory))
+    public SqliteArtifactRegistrationManager(bool inMemory, bool isReadonly = false) : this(new SqliteArtifactContextFactory(inMemory, isReadonly))
     {
     }
 
@@ -45,7 +62,8 @@ public class SqliteArtifactRegistrationManager : EFArtifactRegistrationManager<S
     /// Creates a new instance of <see cref="SqliteArtifactRegistrationManager"/> with the specified Sqlite backing file.
     /// </summary>
     /// <param name="file">Sqlite backing file.</param>
-    public SqliteArtifactRegistrationManager(string file) : this(new SqliteArtifactContextFactory(file))
+    /// <param name="isReadonly">If true, writes to the database are disabled.</param>
+    public SqliteArtifactRegistrationManager(string file, bool isReadonly = false) : this(new SqliteArtifactContextFactory(file, isReadonly))
     {
     }
